tests/test_requirements.txt
Django==2.0
django-progressiveimagefield==0.1.0
Jinja2==2.10
olefile==0.44
Pillow==4.3.0

